+++ CHANGES.rst
@@ -4,6 +4,7 @@ CHANGE HISTORY
 LATEST
 ------
 - Added `allow_int` parameter to `Screen.play()`.
+- Added `max_length` parameter to `Text`.
 - Added support for page up/down in `TextBox`.
 - Added optional scroll bars to `MultiColumnListBox`.
 - Improved performance of double-buffering.